requirements.txt
Django==2.2.10
django_json_field==0.5.7
djangorestframework==3.12.2
